feat: 파일 탐색기 컴포넌트로 분리
