+++ src/app/(main)/note/[id]/graph/page.tsx
@@ -1,5 +1,7 @@
 function Graph() {
-  return <div>그래프 페이지</div>;
+  return (
+    <div className="text-title2 flex h-[696px] w-full items-center justify-center bg-[#D9D9D9]">업데이트 예정</div>
+  );
 }
 
 export default Graph;
