+++ src/app/(main)/note/[id]/memo/page.tsx
@@ -1,5 +1,5 @@
 function Memo() {
-  return <div>메모 페이지</div>;
+  return <div className="h-[696px] w-full bg-[#D9D9D9]">메모 페이지</div>;
 }
 
 export default Memo;
