+++ src/styles/global.css
@@ -8,7 +8,7 @@
 }
 
 @layer components {
-  .text-titel1 {
+  .text-title1 {
     font-family: 'Pretendard Variable';
     font-size: 40px;
     font-style: normal;
